## 内容
- `CUDAExecutionProvider` の `cudnn_conv_algo_search`をデフォルトの`EXHAUSTIVE (0)`から`HEURISTIC (1)`へ変更する
    - https://onnxruntime.ai/docs/execution-providers/CUDA-ExecutionProvider.html#cudnn_conv_algo_search
- これにより、初回実行時にかかる推論時間を大幅に短縮し、`tts()`の`style_id`と`text`を変更した場合に推論時間が大幅に増加(初回実行時と同程度)することを防ぐ
- 以下に示す私の環境でしか実行していません
```text
OS: Microsoft Windows 11 Pro 10.0.26100 64 ビット
CPU: Core i7 12700
GPU: GeForce RTX 3080 Ti
CUDA Version: 12.9
Driver Version: 576.02
```
## 実行結果
<details>
  <summary>依存関係</summary>

```toml
anyhow = "1.0.99"
# 修正前
voicevox_core = { git = "https://github.com/Sanzentyo/voicevox_core.git", branch = "main", features = ["load-onnxruntime"] }
# 修正後
voicevox_core = { git = "https://github.com/Sanzentyo/voicevox_core.git", branch = "perf/use-cudnn-heuristic-algo-search", features = ["load-onnxruntime"] }
```
</details>
<details>
  <summary>実行したコード</summary>

``` rs
use std::time::Instant;

use voicevox_core::{
    StyleId,
    blocking::{Onnxruntime, OpenJtalk, Synthesizer, VoiceModelFile},
};

const VVORT: &str = "./models/voicevox_core/onnxruntime/lib/voicevox_onnxruntime_cuda.dll"; // donwloaderでdeviceをcudaに設定した時にダウンロードされるvoicevox_onnxruntime.dll をrenameしたもの
const OJT_DIC: &str = "models/voicevox_core/dict/open_jtalk_dic_utf_8-1.11";
const MODEL_0_VVM: &str = "./models/voicevox_core/models/vvms/0.vvm";

const ZUNDAMON_NORMAL: StyleId = StyleId(3);
const SHIKOKU_NORMAL: StyleId = StyleId(2);

#[inline(always)]
fn run_tts(
    synth: &Synthesizer<OpenJtalk>,
    text: &str,
    style_id: StyleId,
) -> anyhow::Result<std::time::Duration> {
    let start = Instant::now();
    synth.tts(text, style_id).perform()?;
    Ok(start.elapsed())
}

fn main() -> anyhow::Result<()> {
    let ort = Onnxruntime::load_once().filename(VVORT).perform()?;
    println!("supported providers: {:?}", ort.supported_devices());
    let ojt = OpenJtalk::new(OJT_DIC)?;
    let synth = Synthesizer::builder(ort)
        .text_analyzer(ojt)
        .acceleration_mode(voicevox_core::AccelerationMode::Auto)
        .build()?;

    println!("GPU mode: {}", synth.is_gpu_mode());

    let load_start = Instant::now();
    synth.load_voice_model(&VoiceModelFile::open(MODEL_0_VVM)?)?;
    println!("Loaded {MODEL_0_VVM} in {:?}", load_start.elapsed());

    println!("--- Warmup measurements ---");
    let warmup_text = "ウォームアップテストです";
    let duration_first = run_tts(&synth, warmup_text, ZUNDAMON_NORMAL)?;
    println!(
        "First TTS (style {:?}, text '{}'): {:?}",
        ZUNDAMON_NORMAL, warmup_text, duration_first
    );

    let duration_same_style = run_tts(&synth, warmup_text, ZUNDAMON_NORMAL)?;
    println!("Same style/text after warmup: {:?}", duration_same_style);

    let other_text = "テキストを変更してみます";
    let duration_other_text = run_tts(&synth, other_text, ZUNDAMON_NORMAL)?;
    println!("Same style, different text: {:?}", duration_other_text);

    let duration_other_style = run_tts(&synth, warmup_text, SHIKOKU_NORMAL)?;
    println!(
        "Different style (四国めたん ノーマル): {:?}",
        duration_other_style
    );

    let duration_other_style2 = run_tts(&synth, other_text, SHIKOKU_NORMAL)?;
    println!(
        "Different style/text (四国めたん ノーマル): {:?}",
        duration_other_style2
    );

    let duration_other_style3 = run_tts(&synth, other_text, SHIKOKU_NORMAL)?;
    println!(
        "Same style/text as above (四国めたん ノーマル): {:?}",
        duration_other_style3
    );

    println!("styleを変えた後、元のstyleに戻す");
    let duration_back_to_first_style = run_tts(&synth, warmup_text, ZUNDAMON_NORMAL)?;
    println!(
        "Back to first style/text: {:?}",
        duration_back_to_first_style
    );
    let duration_back_to_first_style2 = run_tts(&synth, "あああ", ZUNDAMON_NORMAL)?;
        println!(
            "Same style, text 'あああ': {:?}",
        duration_back_to_first_style2
    );
    let duration_back_to_first_style3 = run_tts(&synth, "いいい", ZUNDAMON_NORMAL)?;
        println!(
            "Same style, text 'いいい': {:?}",
        duration_back_to_first_style3
    );

    Ok(())
}
```
</details>

<details>
  <summary>修正前</summary>

```text
PS D:\git\signage-backend> cargo run -r -p voicevox-tts --bin warmup_probe_differ
    Updating git repository `https://github.com/Sanzentyo/voicevox_core.git`
     Locking 2 packages to latest compatible versions
      Adding voicevox_core v0.0.0 (https://github.com/Sanzentyo/voicevox_core.git?branch=main#868289d3)
      Adding voicevox_core_macros v0.0.0 (https://github.com/Sanzentyo/voicevox_core.git?branch=main#868289d3)
   Compiling voicevox-tts v0.1.0 (D:\git\signage-backend\voicevox-tts)
    Finished `release` profile [optimized] target(s) in 7.77s
     Running `target\release\warmup_probe_differ.exe`
supported providers: Ok(SupportedDevices { cpu: true, cuda: true, dml: false })
GPU mode: true
Loaded ./models/voicevox_core/models/vvms/0.vvm in 693.5346ms
--- Warmup measurements ---
First TTS (style StyleId(3), text 'ウォームアップテストです'): 4.155918s
Same style/text after warmup: 25.6278ms
Same style, different text: 522.0348ms
Different style (四国めたん ノーマル): 3.9290173s
Different style/text (四国めたん ノーマル): 3.8996648s
Same style/text as above (四国めたん ノーマル): 18.1223ms
styleを変えた後、元のstyleに戻す
Back to first style/text: 20.8202ms
Same style, text 'あああ': 3.8551208s
Same style, text 'いいい': 3.870608s
```
</details>

<details>
  <summary>修正後</summary>

```text
PS D:\git\signage-backend> cargo run -r -p voicevox-tts --bin warmup_probe_differ
    Updating git repository `https://github.com/Sanzentyo/voicevox_core.git`
     Locking 2 packages to latest compatible versions
      Adding voicevox_core v0.0.0 (https://github.com/Sanzentyo/voicevox_core.git?branch=perf%2Fuse-cudnn-heuristic-algo-search#5e2c9450)
      Adding voicevox_core_macros v0.0.0 (https://github.com/Sanzentyo/voicevox_core.git?branch=perf%2Fuse-cudnn-heuristic-algo-search#5e2c9450)
   Compiling voicevox-tts v0.1.0 (D:\git\signage-backend\voicevox-tts)
    Finished `release` profile [optimized] target(s) in 4.39s
     Running `target\release\warmup_probe_differ.exe`
supported providers: Ok(SupportedDevices { cpu: true, cuda: true, dml: false })
GPU mode: true
Loaded ./models/voicevox_core/models/vvms/0.vvm in 715.0655ms
--- Warmup measurements ---
First TTS (style StyleId(3), text 'ウォームアップテストです'): 191.4152ms
Same style/text after warmup: 27.2623ms
Same style, different text: 36.141ms
Different style (四国めたん ノーマル): 29.5294ms
Different style/text (四国めたん ノーマル): 29.9159ms
Same style/text as above (四国めたん ノーマル): 23.4251ms
styleを変えた後、元のstyleに戻す
Back to first style/text: 20.375ms
Same style, text 'あああ': 28.8798ms
Same style, text 'いいい': 20.667ms
```
</details>